requirements.txt
# requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
python-dotenv
